SDCB-12351: New endpoints for account: * /accounts/{accountId}/billing-periods; * /accounts/{accountId}/billing-periods/{billingPeriodId}; * /accounts/{accountId}/billing-periods/{billingPeriodId}/receipt; * /accounts/{accountId}/device-time; * /accounts/{accountId}/device-time-summary;
Removed enpoints:
* /users/{userId}/device-time/reserved
* /users/{userId}/device-time/used
