+++ src/api/APIUserResourceAccount.ts
@@ -1,9 +1,7 @@
 import {APIResource} from './APIResource';
 import {APIResourceUser} from './APIResourceUser';
 import {Account} from './models/Account';
-import {AccountServicePayment} from './models/AccountServicePayment';
-import {BillingPeriodQueryParams} from './models/BillingPeriod';
-import {NoData, NoQueryParams} from './models/HTTP';
+import {NoQueryParams} from './models/HTTP';
 import {VisualTestAccess} from './models/VisualTest';
 
 
@@ -15,18 +13,6 @@ export class APIUserResourceAccount extends APIResource<Account> {
   constructor(parent: APIResourceUser) {
     super(parent);
     this.push('account');
-  }
-
-  // /account-services/{id}/billing-period
-  serviceBillingPeriod(id: number) {
-    if (id == null) {
-      throw new Error('Resource ID cannot be null!');
-    }
-
-    const a = new APIResource<AccountServicePayment, BillingPeriodQueryParams, NoData>(this);
-    a.last += '-services';
-    a.push(id, 'billing-period');
-    return a;
   }
 
   // /account/visualtest/access
